Changed enum values to be all upper case, and moved NodeDimensionality to the correct package
