+++ DataStore/src/main/java/com/rgi/store/routingnetworks/RoutingNetworkStoreReader.java
@@ -27,7 +27,6 @@ package com.rgi.store.routingnetworks;
 import com.rgi.common.BoundingBox;
 import com.rgi.common.Pair;
 import com.rgi.common.coordinate.CoordinateReferenceSystem;
-import com.rgi.store.routingnetworks.osm.NodeDimensionality;
 
 import java.lang.reflect.Type;
 import java.util.List;
