+++ DataStore/src/main/java/com/rgi/store/routingnetworks/triangle/TriangleRoutingNetworkStoreReader.java
@@ -31,7 +31,7 @@ import com.rgi.store.routingnetworks.Edge;
 import com.rgi.store.routingnetworks.Node;
 import com.rgi.store.routingnetworks.RoutingNetworkStoreException;
 import com.rgi.store.routingnetworks.RoutingNetworkStoreReader;
-import com.rgi.store.routingnetworks.osm.NodeDimensionality;
+import com.rgi.store.routingnetworks.NodeDimensionality;
 
 import java.io.File;
 import java.io.IOException;
@@ -294,8 +294,8 @@ public class TriangleRoutingNetworkStoreReader implements RoutingNetworkStoreRea
     public NodeDimensionality getNodeDimensionality() throws RoutingNetworkStoreException
     {
         return this.elevationAttributeIndex < 0
-                                            ? NodeDimensionality.NoElevation
-                                            : NodeDimensionality.HasElevation;
+                                            ? NodeDimensionality.NO_ELEVATION
+                                            : NodeDimensionality.HAS_ELEVATION;
     }
 
     private List<Node> parseNodes() throws RoutingNetworkStoreException
